skip renaming file if no uploaded file

+++ gbdb/models.py
@@ -147,20 +147,21 @@ class ObservationSession(models.Model):
             for field_name, options in rename_files.iteritems():
                 field = getattr(self, field_name)
                 file_name = force_unicode(field)
-                name, ext = os.path.splitext(file_name)
-                keep_ext = options.get('keep_ext', True)
-                final_dest = options['dest']
-                if callable(final_dest):
-                    final_name = final_dest(self, file_name)
-                else:
-                    final_name = os.path.join(final_dest, '%s' % (self.pk,))
-                    if keep_ext:
-                        final_name += ext
-                if file_name != final_name:
-                    field.storage.delete(final_name)
-                    field.storage.save(final_name, field)
-                    field.storage.delete(file_name)
-                    setattr(self, field_name, final_name)
+                if len(file_name):
+                    name, ext = os.path.splitext(file_name)
+                    keep_ext = options.get('keep_ext', True)
+                    final_dest = options['dest']
+                    if callable(final_dest):
+                        final_name = final_dest(self, file_name)
+                    else:
+                        final_name = os.path.join(final_dest, '%s' % (self.pk,))
+                        if keep_ext:
+                            final_name += ext
+                    if file_name != final_name:
+                        field.storage.delete(final_name)
+                        field.storage.save(final_name, field)
+                        field.storage.delete(file_name)
+                        setattr(self, field_name, final_name)
 
         super(ObservationSession,self).save(force_insert=force_insert, force_update=force_update, using=using,
             update_fields=update_fields)
@@ -233,20 +234,21 @@ class BehavioralEvent(MPTTModel):
             for field_name, options in rename_files.iteritems():
                 field = getattr(self, field_name)
                 file_name = force_unicode(field)
-                name, ext = os.path.splitext(file_name)
-                keep_ext = options.get('keep_ext', True)
-                final_dest = options['dest']
-                if callable(final_dest):
-                    final_name = final_dest(self, file_name)
-                else:
-                    final_name = os.path.join(final_dest, '%s' % (self.pk,))
-                    if keep_ext:
-                        final_name += ext
-                if file_name != final_name:
-                    field.storage.delete(final_name)
-                    field.storage.save(final_name, field)
-                    field.storage.delete(file_name)
-                    setattr(self, field_name, final_name)
+                if len(file_name):
+                    name, ext = os.path.splitext(file_name)
+                    keep_ext = options.get('keep_ext', True)
+                    final_dest = options['dest']
+                    if callable(final_dest):
+                        final_name = final_dest(self, file_name)
+                    else:
+                        final_name = os.path.join(final_dest, '%s' % (self.pk,))
+                        if keep_ext:
+                            final_name += ext
+                    if file_name != final_name:
+                        field.storage.delete(final_name)
+                        field.storage.save(final_name, field)
+                        field.storage.delete(file_name)
+                        setattr(self, field_name, final_name)
 
         super(BehavioralEvent,self).save(force_insert=force_insert, force_update=force_update, using=using,
             update_fields=update_fields)
